[step03][test] 경로 탐색 관련 구조 개선에 따른 TC수정
