+++ src/test/java/com/hskim/ladder/model/LineTest.java
@@ -10,7 +10,6 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Stream;
 
@@ -57,7 +56,7 @@ public class LineTest {
     @MethodSource("provideLineAndResult")
     void getStartEndPointMap(Line line, Map<Integer, Integer> expected) {
         // when
-        Map<Integer, Integer> result = line.getStartEndPointMap();
+        Map<Integer, Integer> result = line.getRouteInfo().getRouteMap();
 
         // then
         assertThat(result).isEqualTo(expected);
